refactor: AttackPositions -> AttackRange로 변경
